refactor(30574): export the combining utilities
