+++ hivemq-edge/src/frontend/src/modules/Mappings/combiner/CombinedSchemaLoader.tsx
@@ -2,14 +2,12 @@ import { type FC, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Box, Heading } from '@chakra-ui/react'
 
-import type { DataCombining, DomainTag, TopicFilter } from '@/api/__generated__'
-import { DataIdentifierReference } from '@/api/__generated__'
-import type { DataReference } from '@/api/hooks/useDomainModel/useGetCombinedDataSchemas'
+import type { DataCombining } from '@/api/__generated__'
 import { useGetCombinedDataSchemas } from '@/api/hooks/useDomainModel/useGetCombinedDataSchemas'
 import ErrorMessage from '@/components/ErrorMessage'
 import JsonSchemaBrowser from '@/components/rjsf/MqttTransformation/JsonSchemaBrowser'
 import type { CombinerContext } from '@/modules/Mappings/types'
-import { validateSchemaFromDataURI } from '@/modules/TopicFilters/utils/topic-filter.schema'
+import { getFilteredDataReferences, getSchemasFromReferences } from '@/modules/Mappings/utils/combining.utils'
 
 interface CombinedSchemaLoaderProps {
   formData?: DataCombining
@@ -20,71 +18,15 @@ export const CombinedSchemaLoader: FC<CombinedSchemaLoaderProps> = ({ formData, 
   const { t } = useTranslation()
 
   // TODO[NVL] This is almost a duplicate of the CombinedEntitySelect; reuse
-  const allDataReferences = useMemo(() => {
-    return formContext?.queries?.reduce<DataReference[]>((acc, cur, currentIndex) => {
-      const firstItem = cur.data?.items?.[0]
-      if (!firstItem) return acc
-      if ((firstItem as DomainTag).name) {
-        const tagDataReferences = (cur.data?.items as DomainTag[]).map<DataReference>((tag) => {
-          return {
-            id: tag.name,
-            type: DataIdentifierReference.type.TAG,
-            adapterId: formContext.entities?.[currentIndex]?.id,
-          }
-        })
-        acc.push(...tagDataReferences)
-      } else if ((firstItem as TopicFilter).topicFilter) {
-        const topicFilterDataReferences = (cur.data?.items as TopicFilter[]).map<DataReference>((topicFilter) => ({
-          id: topicFilter.topicFilter,
-          type: DataIdentifierReference.type.TOPIC_FILTER,
-          adapterId: undefined,
-        }))
-        acc.push(...topicFilterDataReferences)
-      }
+  const references = useMemo(() => {
+    return getFilteredDataReferences(formData, formContext)
+  }, [formContext, formData])
 
-      return acc
-    }, [])
-  }, [formContext?.entities, formContext?.queries])
-
-  const allSchemas = useMemo(() => {
-    const tags = formData?.sources?.tags || []
-    const topicFilters = formData?.sources?.topicFilters || []
-    const indexes = [...tags, ...topicFilters]
-
-    const selectedReferences = allDataReferences?.filter((dataReference) => indexes.includes(dataReference.id)) || []
-    return selectedReferences.reduce<DataReference[]>((acc, current) => {
-      const isAlreadyIn = acc.find((item) => item.id === current.id && item.type === current.type)
-      if (!isAlreadyIn) {
-        return acc.concat([current])
-      }
-      return acc
-    }, [])
-  }, [allDataReferences, formData?.sources?.tags, formData?.sources?.topicFilters])
-
-  const queries = useGetCombinedDataSchemas(allSchemas)
+  const schemaQueries = useGetCombinedDataSchemas(references)
 
   const displayedSchemas = useMemo(() => {
-    return allSchemas.map((dataReference, index) => {
-      const { data } = queries[index]
-
-      // TODO[30744] Type of schema inconsistent between tag and topic filter
-      if (typeof data === 'string') {
-        dataReference.schema = validateSchemaFromDataURI(data)
-      } else if (typeof data === 'object') {
-        dataReference.schema = {
-          schema: data || undefined,
-          status: 'success',
-          message: t('topicFilter.schema.status.success'),
-        }
-      } else {
-        dataReference.schema = {
-          status: 'warning',
-          message: t('topicFilter.error.schema.noAssignedSchema', { context: dataReference.type }),
-        }
-      }
-      return dataReference
-    })
-  }, [allSchemas, queries, t])
+    return getSchemasFromReferences(references, schemaQueries)
+  }, [references, schemaQueries])
 
   return (
     <Box borderWidth={1} p={3}>
